Remove check to make sure that paragraphs start at a new line
This is necessary to correctly render list items for both ordered and
unordered lists. Text within list items is wrapped in a paragraph, so
ensuring a newline would insert a newline in between the bullet or
number in the list and the text.

Paragraphs should still have a newline enforced after their text though.