Support serializing ordered and unordered lists
This adds the concept of a `prefix` to the WikiTextSerializerState,
which contains a string to prefix each paragraph type with. This is
helpful in serializing nested lists, because for each depth of list, the
prefix can contain the necessary wiki syntax to place before each list
item in an ordered or unordered list.

This also adds a new test which makes sure that the prefix can return to
it's previous state after a lower level of list nesting is serialized.