Fix bug for nested lists of different types
This fixes a bug where nested lists of ordered and bullet types are not
rendered correctly. This is because the prefix cannot be built
recursively, otherwise we would end up with something like:

\# Ordered entry
\#* Nested bullet entry

The new functionality is to change the prefix to contain the correct
number of nested symbol characters. This generates the correct syntax of:

\# Ordered entry
** Nested bullet entry

This also includes a new test for asserting that this functionality
works correctly.