Add link support to the serializer
This add links to the minimal and standard schema, and a new mark rule
for the serializer. It also adds two new tests for links.

Links can contain hrefs, titles, and target attributes. iFixit wiki
syntax does not support title attributes, but does support
`target="_blank"` attributes which open the linked page in a new tab.

A link of the form:

<a href="https://ifixit.com">iFixit</a>

will look like

[https://ifixit.com|iFixit]

while a link like

<a href="https://ifixit.com" target="_blank">iFixit</a>

will be serialized to

[https://ifixit.com|iFixit|new_window=true]