Update prosemirror-model version requirement, add schema-list
This bumps the prosemirror-model version requiement to be equal with
upstream, and adds the prosemirror-schema-list module as well, since the
list tests require it.

This also most of the module imports in test functions, which were
previously relying on relative paths to other prosemirror modules. Now
they reference the module name directly.