Close inline marks in the correct order
There was a bug where nested marks where not being closed in the correct
order. If a user enterered text that was underlined, and then subscript
and underlined, the serializer would close the underlined mark before the
subscript mark, which is invalid wiki text.

The previous behvaior used Array.prototype.reverse, which mutated the
existing array and led to some weird behavior where a single array was
being reversed more than once.

This modifies the behavior so that closeMarks always does a reduceRight
to close tags in reverse order, and openMarks always does a reduce to
open tags in the original order.